Add optional parameter to not checkpoint failed job executions.
Until now, the default behaviour was to always checkpoint all received messages, regardless of success. The rationale for this is to make sure that the eventhub does not block, and the function is responsible for handling exceptions.

However, there are use cases where this is not required behaviour.
- In some cases order needs to be guaranteed. If the function is responsible for putting failed events in a poison queue, then this will break the order guarantee. It will continue processing new events, and can't process the poisoned event later (due to order contraints). If the function blocks on the failed message, it will keep the order guarantees.
- The code that puts a failed event in a poison queue can fail as well. If this happens, the data is lost in the current situation. With a non-checkpointing function it will keep the event in the queue until it is processed correctly.

The EventHubTrigger now has an additional paramter 'CheckpointOnFailure'. If this parameter is set to True (the default value) it will checkpoint all events, regardless of success or failure. If this parameter is False, it will not checkpoint events that caused exceptions.